data-xfer: Adding dependency/mock framework to ObjectWrapper
Since Custom Metadata records can not be created/edited in Apex code, we can't rely on creating those records for our tests. Based on a couple internet articles I found I created a system that uses dependency injection to provide the class that actually retrieves the records, and this class can be mocked.

I have realized that I do not have a query method on my selector, and the way I have configured it is probably not going to work. I may end up using more of a passthrough, which I was hoping to avoid because it seems like not as clean of an implementation.