data-xfer: Added query method to metadata selector class
I added a query() method to the metadata selector. It is up to the subclass to figure out an implementation, but for now the datasource (the only subclass so far) just acts as a passthrough, returning a list of records that was passed into the constructor.

I don't think this solution is particularly elegant, but it will let me move ahead for now.