data-xfer: Created a test class for ServiceHandler
I created my first test class, for the ServiceHandler class. There are a lot of moving parts that should be reusable in the future. The first is creating a mock framework for loading custom metadata records, though the bulk of this work was in a previous commit. The second is a mock framework for the ServiceHandler, so we can customize and test specific aspects of the ServiceHandler class.

For now there are two test methods that test how request data is serialized into URL parameters for GET requests.