data-xfer: Added a Listener pattern for parsing Configuration object
The Configuration object is structured to have multiple trees. Instead of having every piece of code that needs to interpret the configuration implement a traversal of one of those trees, I've created a Listener interface (actually multiple interfaces so the implementor can select only the relevant events) for processing the object trees.