data-xfer: Added subrelations to ObjectRelation class
I realized that the Object_Relation__mdt metadata type is inadequate to store junction objects since a junction object really needs to store 4 fields and the existing relation object can only hold 2. So I made the junction object just a parent object of two other junction objects. The junction object specifies that it is of type "junction", that it relates the two types of objects, what type the intermediate object is, as well as the two child relations (the names are stored in the Field1__c and Field2__c fields). The child objects then act as normal reference relations between the junction object and the objects on either side of the junction.