data-xfer: Changed ObjectRelation.IntermediateObject from String to Lookup field
There was no reason for this to be a String. I was thinking if someone wants to use a custom object as a junction object it might be handy to have this as a string, but no custom metadata records will be created after install (or package creation, really) so we'll need another mechanism for that anyway.