data-xfer: Removed DataTransfer.Relation class
In my attempt to have few relation classes, I have removed the DataTransfer.Relation class. The TargetRecordHandle.ParentRelation class is now used, and it is constructed in pieces: first it is created with an object path, a parent field, and a child field, then later the parent object can be set (and is validated against the path passed to the constructor). An instance must have the parent object set to be added as a relation for another TargetRecordHandle instance.