data-xfer: Added Configuration.sourceObjects and .targetObjects members
I realized that the Configuration.Listener framework is not buying me anything. All I really need is a flat list of all the mapped objects (with path and parent populated for each record) to do the same thing. What are now listeners can instead just iterate over that list to get all the same information. So I have updated the Configuration class to create this list so it can be used by other classes in the future (I have not yet gotten rid of the existing Listener implementations).

In addition, I updated the parseTreeRecursive method to use the new RecordPath class instead of just using a String. I will need to keep expanding out the use of RecordPath but this is a first step.

I also added a ConfigurationTest class to ensure the configuration parsing is working as expected.