data-xfer: Created scripts to convert JSON <=> Apex
Since Apex doesn't have a great solution for multiline strings I was having trouble changing JSON to Apex code and vice versa (e.g. figuring out what is wrong with a given JSON string in Apex code). So I used ChatGPT to create two bash scripts to accept input via stdin and print convert the input to the other format. It was hard getting the formatting correct (without installing the jq command), so when JSON is created from an Apex code snippet, it allows the user to load a webpage that automatically displays the formatted json text.