data-xfer: Updated DataRecord class
Previously the DataRecord class was primarily meant to handle marshalling and unmarshalling data based on the formats required for each service. I plan to move this functionality into a new DataMarshaller class in the near future, but for now I've added new methods to DataRecord for interpreting and traversing the record that is passed into it.

I've also removed the getRecords() method from the handler base class since this was meant to be used for instantiating service-specific subclasses of DataRecord, which is no longer the intended use. It will be eventually replaced with a getDataMarshaller() method.