data-xfer: Implemented DataTransfer.loadDataRecords() method
Created a method to parse the data returned from a service, wrapping parts of the tree in DataRecord objects based on the source objects parsed out of the Configuration object. These DataRecord objects are then added to a map, with the key being the path to that object in the tree. This map will then be used to find all of the pieces of data that need to be used for generating the new objects.