data-xfer: Updated Configuration.MappedObject to store a RecordPath
Changed MappedObject.path from a string to a RecordPath. This is now the main place path objects are created, so this touched a lot of other classes. All tests are passing, but this change made me realize that since we are storing paths in other objects as well as using them as map keys, they really need to be immutable. There was one tricky bug already getting the tests to pass, and allowing these objects to be mutated is just going to be inviting more trouble in the future. The change will be in a future commit, though.