data-xfer: Made RecordPath instances immutable
Since RecordPath instances are used as map keys and are stored in other objects, it was becoming difficult to make sure that changes to an instance wouldn't have unintended side effects. Before I was doing a lot of copies to get around the problem, but the real strategy is to make the objects immutable, so that's what I did. Now all methods that make a change to the path return a new instance with the changes. I put in protections as well (in the property setters) to try to ensure that the member variables are not changed after instantiation.